Part of WFOF-299. Update LTV/CAC view to improve acquisition and period logic
Refines acquisition date calculation to group by region and adjusts the logic for determining 'New' customers to use the start of the current month. Updates prior period selection to use the last day of the previous month and adds a new subquery to count gross acquisitions in the current month by region. Also updates marketing spend and join logic for more accurate reporting.